refactor(worker,shared): share integration helpers across export integrations
Three behaviour-preserving extractions, so a second export integration can
reuse what blob storage already had rather than copying it:

- Move the deterministic customer-fault classifier out of the blob-storage
  feature into a neutral integrations module. Blob keeps its own
  disable-metric constant and now re-exports the classifier, so its imports
  and behaviour are unchanged.
- Extract the Prisma P2025 "record deleted mid-run" predicate, which was a
  private copy inside the blob handler, into the same neutral module.
- Parameterise the export-failure email. The sender mechanics (SMTP guards,
  subject sanitisation, CRM bcc, error logging) and the React Email template
  now exist once and take the integration's labels and copy as arguments;
  blob storage becomes a thin caller supplying its own wording.

The email refactor keeps blob's rendered output identical: same subject
lines, same visible body text, same links. Only React's zero-width text-node
separators shift, because literal labels became interpolated values.

Three separate label casings are passed rather than derived, since not every
integration name lower-cases and title-cases predictably.
